+++ frontend/src/main.tsx
@@ -4,9 +4,10 @@ import "@fontsource/jetbrains-mono/latin-400.css";
 import { sourceContentCss } from "./components/dashboard/SourceContent";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import App from "./App";
+import {DraftProvider} from "./contexts/DraftContext";
 import { ThemeProvider } from "./contexts/ThemeContext";
 import { SettingsProvider } from "./contexts/SettingsContext";
 import "./index.css";
@@ -17,14 +18,14 @@ const queryClient = new QueryClient({
   },
 });
 
+const router=createBrowserRouter([{path:"*",element:<DraftProvider><App/></DraftProvider>}]);
+
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <ThemeProvider>
         <SettingsProvider>
-          <BrowserRouter>
-            <App />
-          </BrowserRouter>
+          <RouterProvider router={router} />
         </SettingsProvider>
       </ThemeProvider>
     </QueryClientProvider>